+++ src/features/ArticleCardList/presentations/index.tsx
@@ -1,0 +1,23 @@
+import React from "react";
+import {
+  ArticleCardListItemPresentation,
+  Props as ArticleCardListItemPresentationProps,
+} from "./ArticleCardListItem";
+
+type Item = ArticleCardListItemPresentationProps & {
+  id: string;
+};
+
+type Props = {
+  items: Item[];
+};
+
+export const ArticleCardListPresentation: React.FC<Props> = ({ items }) => {
+  return (
+    <div>
+      {items.map((item) => (
+        <ArticleCardListItemPresentation key={item.id} {...item} />
+      ))}
+    </div>
+  );
+};
